chore: add all end points relative to manager service

+++ src/main/java/com/airbnl/dbcontrollerservice/model/Hotel.java
@@ -18,6 +18,6 @@ public class Hotel {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private long id;
     private String name;
-    private int managerId;
-    private int countryId;
+    private long managerId;
+    private long countryId;
 }
